feat: separate explore page from landing page
