+++ frontend/src/App.tsx
@@ -1,6 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
 import { Layout } from './components/Layout';
 import { HomePage } from './pages/Home';
+import { ExplorePage } from './pages/Explore';
 import { CampaignPage } from './pages/Campaign';
 import { CreatePage } from './pages/Create';
 import { MyCampaignsPage } from './pages/MyCampaigns';
@@ -10,6 +11,7 @@ function App() {
     <Layout>
       <Routes>
         <Route path="/" element={<HomePage />} />
+        <Route path="/explore" element={<ExplorePage />} />
         <Route path="/campaign/:id" element={<CampaignPage />} />
         <Route path="/create" element={<CreatePage />} />
         <Route path="/my-campaigns" element={<MyCampaignsPage />} />
