+++ frontend/src/pages/Home.tsx
@@ -1,55 +1,28 @@
 import { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { CampaignCard, type Campaign } from '../components/CampaignCard';
-import { Search, Filter, Plus, Zap, Globe, Shield, RefreshCw, ArrowRight, Wallet, CheckCircle2 } from 'lucide-react';
-import { CAMPAIGN_CATEGORIES } from '../lib/ipfs';
-import { getAllCampaigns, filterCampaigns, type FullCampaign } from '../lib/campaigns';
+import { Plus, Zap, Globe, Shield, RefreshCw, ArrowRight, Wallet, CheckCircle2 } from 'lucide-react';
+import { getAllCampaigns, type FullCampaign } from '../lib/campaigns';
 
 export function HomePage() {
   const [campaigns, setCampaigns] = useState<FullCampaign[]>([]);
-  const [searchQuery, setSearchQuery] = useState('');
-  const [selectedCategory, setSelectedCategory] = useState<string>('All');
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+  const [, setLoading] = useState(true);
 
-  // Fetch campaigns on mount
+  // Fetch campaigns on mount for stats
   useEffect(() => {
     loadCampaigns();
   }, []);
 
   const loadCampaigns = async () => {
     setLoading(true);
-    setError(null);
     try {
       const data = await getAllCampaigns();
       setCampaigns(data);
     } catch (err) {
       console.error('Failed to load campaigns:', err);
-      setError('Failed to load campaigns. Please try again.');
     } finally {
       setLoading(false);
     }
   };
-
-  // Filter campaigns
-  const filteredCampaigns = filterCampaigns(campaigns, {
-    searchQuery,
-    category: selectedCategory,
-  });
-
-  // Convert FullCampaign to Campaign for CampaignCard
-  const displayCampaigns: Campaign[] = filteredCampaigns.map(c => ({
-    id: c.id,
-    title: c.title,
-    description: c.description,
-    imageUrl: c.imageUrl,
-    goal: c.goal,
-    raised: c.raised,
-    backers: c.backers,
-    daysLeft: c.daysLeft,
-    category: c.category,
-    owner: c.owner,
-  }));
 
   // Stats
   const totalRaised = campaigns.reduce((sum, c) => sum + c.raised, 0);
@@ -93,13 +66,13 @@ export function HomePage() {
               <Plus className="w-5 h-5" />
               Start a Campaign
             </Link>
-            <a 
-              href="#campaigns" 
+            <Link 
+              to="/explore" 
               className="btn-secondary flex items-center gap-2 text-lg px-8 py-4"
             >
               Explore Projects
               <ArrowRight className="w-5 h-5" />
-            </a>
+            </Link>
           </div>
 
           {/* Trust Indicators */}
@@ -221,7 +194,7 @@ export function HomePage() {
       </section>
 
       {/* Stats Section */}
-      <section className="grid grid-cols-1 sm:grid-cols-3 gap-6 max-w-4xl mx-auto px-4">
+      <section className="grid grid-cols-1 sm:grid-cols-3 gap-6 max-w-4xl mx-auto px-4 pb-20">
         <div className="stat-card group hover:border-primary-500/30 transition-colors">
           <div className="w-12 h-12 mx-auto mb-3 rounded-xl bg-primary-500/10 flex items-center justify-center group-hover:scale-110 transition-transform">
             <Wallet className="w-6 h-6 text-primary-400" />
@@ -249,109 +222,6 @@ export function HomePage() {
           </div>
           <div className="stat-label">Funded Projects</div>
         </div>
-      </section>
-
-      {/* Campaigns Section */}
-      <section id="campaigns" className="max-w-7xl mx-auto px-4 space-y-8">
-        <div className="flex flex-col md:flex-row items-end justify-between gap-6">
-          <div>
-            <h2 className="text-3xl font-heading font-bold text-dark-100 mb-2">
-              Explore Campaigns
-            </h2>
-            <p className="text-dark-400">
-              Discover and fund the next generation of Bitcoin applications
-            </p>
-          </div>
-          
-          {/* Filter Bar */}
-          <div className="flex flex-col sm:flex-row gap-4 w-full md:w-auto">
-            <div className="relative flex-1 sm:w-64">
-              <Search className="absolute left-4 top-1/2 -translate-y-1/2 w-4 h-4 text-dark-500" />
-              <input
-                type="text"
-                placeholder="Search..."
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                className="input pl-10 py-2 text-sm"
-              />
-            </div>
-            <button
-              onClick={loadCampaigns}
-              disabled={loading}
-              className="btn-secondary py-2 px-4 flex items-center justify-center gap-2"
-            >
-              <RefreshCw className={`w-4 h-4 ${loading ? 'animate-spin' : ''}`} />
-              <span className="hidden sm:inline">Refresh</span>
-            </button>
-          </div>
-        </div>
-
-        {/* Categories */}
-        <div className="flex items-center gap-2 overflow-x-auto pb-4 scrollbar-hide -mx-4 px-4 md:mx-0 md:px-0">
-          <Filter className="w-4 h-4 text-dark-500 flex-shrink-0 mr-2" />
-          {['All', ...CAMPAIGN_CATEGORIES].map((category) => (
-            <button
-              key={category}
-              onClick={() => setSelectedCategory(category)}
-              className={`px-4 py-1.5 rounded-full text-sm font-medium whitespace-nowrap transition-all ${
-                selectedCategory === category
-                  ? 'bg-primary-500 text-dark-900 shadow-lg shadow-primary-500/20'
-                  : 'bg-white/5 text-dark-300 hover:bg-white/10 hover:text-dark-100 border border-white/5'
-              }`}
-            >
-              {category}
-            </button>
-          ))}
-        </div>
-
-        {/* Grid */}
-        {loading ? (
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
-            {[1, 2, 3, 4, 5, 6].map((i) => (
-              <div key={i} className="glass-card overflow-hidden animate-pulse border-white/5">
-                <div className="h-56 bg-dark-800" />
-                <div className="p-6 space-y-4">
-                  <div className="h-6 bg-dark-800 rounded w-3/4" />
-                  <div className="h-4 bg-dark-800 rounded w-full" />
-                  <div className="h-4 bg-dark-800 rounded w-2/3" />
-                  <div className="pt-4 flex justify-between">
-                    <div className="h-8 bg-dark-800 rounded w-20" />
-                    <div className="h-8 bg-dark-800 rounded w-20" />
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        ) : error ? (
-          <div className="glass-card text-center py-20 border-red-500/20">
-            <p className="text-dark-400 text-lg mb-6">{error}</p>
-            <button onClick={loadCampaigns} className="btn-primary">
-              Try Again
-            </button>
-          </div>
-        ) : displayCampaigns.length === 0 ? (
-          <div className="glass-card text-center py-24 border-white/5">
-            <div className="w-16 h-16 bg-dark-800 rounded-full flex items-center justify-center mx-auto mb-6">
-              <Search className="w-8 h-8 text-dark-600" />
-            </div>
-            <h3 className="text-xl font-heading font-semibold text-dark-200 mb-2">No campaigns found</h3>
-            <p className="text-dark-400 text-lg mb-8 max-w-md mx-auto">
-              {campaigns.length === 0 
-                ? 'Be the first to launch a campaign on Impact-X!'
-                : 'We couldn\'t find any campaigns matching your criteria.'}
-            </p>
-            <Link to="/create" className="btn-primary inline-flex items-center gap-2">
-              <Plus className="w-5 h-5" />
-              Start a Campaign
-            </Link>
-          </div>
-        ) : (
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
-            {displayCampaigns.map((campaign) => (
-              <CampaignCard key={campaign.id} campaign={campaign} />
-            ))}
-          </div>
-        )}
       </section>
     </div>
   );
